basic integration test working
