+++ legion.cabal
@@ -13,55 +13,72 @@ cabal-version:       >=1.10
 
 library
   hs-source-dirs:      src
-  exposed-modules:     Lib
+  exposed-modules:     Lib, Server
   default-language:    Haskell2010
-  build-depends:       base >= 4.7 && < 5
+  build-depends:       GenericPretty
+                     , HTTP >= 4000.3.6
+                     , Spock >= 0.11
                      , aeson
+                     , async
                      , base >= 4.7 && < 5
+                     , binary
                      , bytestring
                      , cryptohash
+                     , distributed-process
+                     , distributed-process-p2p
+                     , either
+                     , hslogger
                      , mtl
+                     , tasty >= 0.11.2
+                     , tasty-hunit >= 0.9.2
                      , text
-                     , binary
-                     , GenericPretty
                      , time
 
 executable legion-exe
   hs-source-dirs:      app
   main-is:             Main.hs
   default-language:    Haskell2010
-  build-depends:       Spock >= 0.11
+  build-depends:       GenericPretty
+                     , HTTP >= 4000.3.6
+                     , Spock >= 0.11
                      , aeson
+                     , async
                      , base >= 4.7 && < 5
+                     , binary
                      , bytestring
                      , cryptohash
+                     , distributed-process
+                     , distributed-process-p2p
+                     , either
+                     , hslogger
+                     , legion
                      , mtl
+                     , tasty >= 0.11.2
+                     , tasty-hunit >= 0.9.2
                      , text
                      , time
-                     , distributed-process-p2p
-                     , distributed-process
-                     , async
-                     , either
-                     , hslogger 
-                     , binary
-                     , GenericPretty
-                     , legion
 
 test-suite legion-test
   type:                exitcode-stdio-1.0
   hs-source-dirs:      app test
   main-is:             Spec.hs
   default-language:    Haskell2010
-  build-depends:  Spock >= 0.11
-                , base
-                , tasty
-                , tasty-hunit >= 0.9.2
+  build-depends:  HTTP >= 4000.3.6
+                , HTTP >= 4000.3.6
+                , Spock >= 0.11
                 , aeson
+                , async
+                , base
                 , base >= 4.7 && < 5
                 , bytestring
                 , cryptohash
+                , legion
                 , mtl
+                , tasty >= 0.11.2
+                , tasty >= 0.11.2
+                , tasty >= 0.11.2
+                , tasty-hunit >= 0.9.2
+                , tasty-hunit >= 0.9.2
                 , text
                 , time
-                , legion
